+++ src/graphics/material.rs
@@ -4,18 +4,18 @@ use std::hash::Hash;
 #[derive(Copy, Clone, Eq, PartialEq, bytemuck::Pod, bytemuck::Zeroable, Hash, Debug)]
 pub struct Material {
     // TODO: check alignment.
-    base_color_texture_array_id: usize,
-    base_color_texture_id: usize,
-    normal_texture_array_id: usize,
-    normal_texture_id: usize,
+    base_color_texture_array_id: u32,
+    base_color_texture_id: u32,
+    normal_texture_array_id: u32,
+    normal_texture_id: u32,
 }
 
 impl Material {
     pub fn new(
-        base_color_texture_array_id: usize,
-        base_color_texture_id: usize,
-        normal_texture_array_id: usize,
-        normal_texture_id: usize,
+        base_color_texture_array_id: u32,
+        base_color_texture_id: u32,
+        normal_texture_array_id: u32,
+        normal_texture_id: u32,
     ) -> Self {
         Self {
             base_color_texture_array_id,
